Rename `revision` package to `revisions` to help avoid name conflicts.
