+++ src/internal/revisions/closed.go
@@ -1,4 +1,4 @@
-package revision
+package revisions
 
 import (
 	"context"
