+++ src/rinqamqp/internal/notifyamqp/factory.go
@@ -3,7 +3,7 @@ package notifyamqp
 import (
 	"github.com/rinq/rinq-go/src/internal/localsession"
 	"github.com/rinq/rinq-go/src/internal/notify"
-	"github.com/rinq/rinq-go/src/internal/revision"
+	"github.com/rinq/rinq-go/src/internal/revisions"
 	"github.com/rinq/rinq-go/src/rinq/ident"
 	"github.com/rinq/rinq-go/src/rinq/options"
 	"github.com/rinq/rinq-go/src/rinqamqp/internal/amqputil"
@@ -14,7 +14,7 @@ func New(
 	peerID ident.PeerID,
 	opts options.Options,
 	sessions localsession.Store,
-	revisions revision.Store,
+	revs revisions.Store,
 	channels amqputil.ChannelPool,
 ) (notify.Notifier, notify.Listener, error) {
 	channel, err := channels.GetQOS(opts.SessionWorkers) // do not return to pool, use for listener
@@ -30,7 +30,7 @@ func New(
 		peerID,
 		opts.SessionWorkers,
 		sessions,
-		revisions,
+		revs,
 		channel,
 		opts.Logger,
 		opts.Tracer,
